refactor: sperate util files
