+++ packages/oruga-next/src/components/modal/index.ts
@@ -9,7 +9,7 @@ import {
 } from "@/utils/plugins";
 
 /** export modal specific types */
-export type { ModalProps } from "./types";
+export type { ModalProps } from "./ModalProgrammatic";
 
 /** export modal plugin */
 export default {
