+++ packages/oruga-next/src/components/notification/NotificationProgrammatic.ts
@@ -1,6 +1,8 @@
 import { createVNode, render } from "vue";
+import type { ComponentProps } from "vue-component-type-helpers";
 
 import NotificationNotice from "./NotificationNotice.vue";
+import Notification from "./Notification.vue";
 
 import InstanceRegistry from "@/utils/InstanceRegistry";
 import { VueInstance } from "@/utils/plugins";
@@ -11,13 +13,16 @@ import type {
     ProgrammaticExpose,
     ProgrammaticInstance,
 } from "@/types";
-import type { NotifcationNoticeProps, NotifcationProps } from "./types";
 
 declare module "../../index" {
     interface OrugaProgrammatic {
         notification: typeof NotificationProgrammatic;
     }
 }
+
+export type NotifcationProps = ComponentProps<typeof Notification>;
+
+export type NotifcationNoticeProps = ComponentProps<typeof NotificationNotice>;
 
 const instances = new InstanceRegistry<typeof NotificationNotice>();
