+++ packages/oruga-next/src/components/notification/types.ts
@@ -1,12 +1,4 @@
 import type { ClassDefinition, ComponentConfigBase } from "@/types";
-import Notification from "./Notification.vue";
-import NotificationNotice from "./NotificationNotice.vue";
-
-export type NotifcationProps = InstanceType<typeof Notification>["$props"];
-
-export type NotifcationNoticeProps = InstanceType<
-    typeof NotificationNotice
->["$props"];
 
 declare module "../../index" {
     interface OrugaOptions {
