+++ packages/oruga-next/src/components/select/types.ts
@@ -1,18 +1,5 @@
 import type { ClassDefinition, ComponentConfigBase } from "@/types";
 
-/**
- * Options should always be formatted as an array of objects with label and value
- * properties.
- */
-export interface OptionsItem<V = unknown> {
-    label: string;
-    value: V;
-    attrs?: {
-        disabled?: boolean;
-    } & Record<string, any>;
-    [index: string]: any;
-}
-
 declare module "../../index" {
     interface OrugaOptions {
         select?: ComponentConfigBase &
