+++ packages/oruga-next/src/components/select/utils.ts
@@ -1,0 +1,12 @@
+/**
+ * Options should always be formatted as an array of objects with label and value
+ * properties.
+ */
+export interface OptionsItem<V = unknown> {
+    label: string;
+    value: V;
+    attrs?: {
+        disabled?: boolean;
+    } & Record<string, any>;
+    [index: string]: any;
+}
