+++ packages/oruga-next/src/components/sidebar/SidebarProgrammatic.ts
@@ -1,7 +1,7 @@
 import { createVNode, render } from "vue";
+import type { ComponentProps } from "vue-component-type-helpers";
 
 import Sidebar from "./Sidebar.vue";
-import type { SidebarProps } from "./types";
 
 import InstanceRegistry from "@/utils/InstanceRegistry";
 import { VueInstance } from "@/utils/plugins";
@@ -13,6 +13,8 @@ declare module "../../index" {
         sidebar: typeof SidebarProgrammatic;
     }
 }
+
+export type SidebarProps = ComponentProps<typeof Sidebar>;
 
 const instances = new InstanceRegistry<typeof Sidebar>();
