+++ packages/oruga-next/src/components/sidebar/types.ts
@@ -1,7 +1,4 @@
 import type { ClassDefinition, ComponentConfigBase } from "@/types";
-import Sidebar from "./Sidebar.vue";
-
-export type SidebarProps = InstanceType<typeof Sidebar>["$props"];
 
 declare module "../../index" {
     interface OrugaOptions {
