+++ packages/oruga-next/src/components/slider/types.ts
@@ -1,12 +1,4 @@
 import type { ClassDefinition, ComponentConfigBase } from "@/types";
-import Slider from "./Slider.vue";
-
-export type SliderProps = InstanceType<typeof Slider>["$props"];
-
-export type SliderComponent = {
-    min: number;
-    max: number;
-};
 
 declare module "../../index" {
     interface OrugaOptions {
