+++ packages/oruga-next/src/components/steps/index.ts
@@ -6,7 +6,7 @@ import StepItem from "./StepItem.vue";
 import { registerComponent } from "@/utils/plugins";
 
 /** export steps specific types */
-export type * from "./types";
+export type * from "./utils";
 
 /** export steps plugin */
 export default {
