+++ packages/oruga-next/src/components/tabs/index.ts
@@ -6,7 +6,7 @@ import TabItem from "./TabItem.vue";
 import { registerComponent } from "@/utils/plugins";
 
 /** export tabs specific types */
-export type * from "./types";
+export type * from "./utils";
 
 /** export tabs plugin */
 export default {
