+++ packages/oruga-next/src/components/tabs/types.ts
@@ -1,32 +1,8 @@
-import type { ProviderItem } from "@/composables";
-import TabItem from "./TabItem.vue";
 import type {
-    ClassBind,
     ClassDefinition,
     ComponentConfigBase,
     DynamicComponent,
 } from "@/types";
-
-export type TabItemProps = InstanceType<typeof TabItem>["$props"];
-
-export type TabItemSlots = InstanceType<typeof TabItem>["$slots"];
-
-export type TabItemComponent = TabItemProps & {
-    $slots: TabItemSlots;
-    headerIconClasses: ClassBind[];
-    headerTextClasses: ClassBind[];
-    isTransitioning: boolean;
-    activate: (index: number) => void;
-    deactivate: (index: number) => void;
-};
-
-export type TabsComponent = {
-    activeId: number;
-    type: string;
-    vertical: boolean;
-};
-
-export type TabItem = Omit<ProviderItem, "data"> & TabItemComponent;
 
 declare module "../../index" {
     interface OrugaOptions {
